I.In order to access my website, I need to sign up and then log in.
During my signup, I provide my username, full name, password, sex, and date of birth.
The data I enter is stored in a json file.

When I attempt to log in, the application reads my user data from the text file and searches for a match.

II. The gallery is designed to be dynamic. There is a gallery json file containing a list of image Urls.

To add an image to the gallery, I save it in the gallery folder and append its name to the gallery file.

The gallery is loaded dynamically based on the entries in the gallery file.

III. On each page of my application, a welcome message appears in the top right corner, displaying my name and a logout button.
Access to the pages is restricted unless I have successfully logged in.